requirements.txt
python-dotenv>=1.0.1
Pillow>=10.4.0
mss>=9.0.1
openai>=1.51.0
requests>=2.32.3
pytesseract>=0.3.13
easyocr>=1.7.2
edge-tts>=7.0.0
pyttsx3>=2.98
